+++ src/main/java/az/edu/turing/ProblemOfabsum.java
@@ -1,19 +1,17 @@
 package az.edu.turing;
 
 import java.util.Scanner;
+
 public class ProblemOfabsum {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         for (int i = 0; i < 2; i++) {
 
-            int a = sc.nextInt();
-            int b = sc.nextInt();
+            long  a = sc.nextLong();
+            long b = sc.nextLong();
             if (Math.abs(a) <= 30000 && Math.abs(b) <= 30000) {
-
-                System.out.println(a+b);
-
+                System.out.println(a + b);
             }
         }
-
     }
 }
